requirements.txt
torch>=2.0.0
transformers>=4.36.0
peft>=0.6.0
huggingface_hub>=0.19.0
tokenizers>=0.15.0
accelerate>=0.24.0
python-telegram-bot==21.0.1
sentencepiece
protobuf
bitsandbytes
